refactor: move types into separate folder
