+++ src/resources/tasks/task.memory.repository.ts
@@ -4,37 +4,37 @@ let tasks: TaskClass[] = [];
 
 export class TasksRepo {
   /**
- * Get all tasks
- * @returns \{Promise\} Promise object represents collection of tasks
- */
+   * Get all tasks
+   * @returns \{Promise\} Promise object represents collection of tasks
+   */
   async getAll() {
     return tasks;
   }
 
   /**
- * Get all filtered tasks
- * @param filter - object with key to search and value to compare
- * @returns \{Promise\} Promise objects represents all tasks corresponding to the filter
- */
+   * Get all filtered tasks
+   * @param filter - object with key to search and value to compare
+   * @returns \{Promise\} Promise objects represents all tasks corresponding to the filter
+   */
   async search({ key, value }: { key: keyof TaskClass, value: string }) {
     const filteredCollection = tasks.filter((elem: TaskClass) => elem[key] === value);
     return filteredCollection;
   }
 
   /**
- * Get one task by id
- * @param id - task id string
- * @returns \{Promise\} Promise object represents task with a passed id
- */
+   * Get one task by id
+   * @param id - task id string
+   * @returns \{Promise\} Promise object represents task with a passed id
+   */
   async getOne(id: string) {
     return tasks.find((elem) => elem.id === id);
   }
 
   /**
- * Create one task
- * @param newItem - object with task parameters;
- * @returns \{Promise\} Promise object represents newly create task
- */
+   * Create one task
+   * @param newItem - object with task parameters;
+   * @returns \{Promise\} Promise object represents newly create task
+   */
   async createOne(newItem: ITaskToCreate) {
     const item = new TaskClass(newItem);
 
@@ -43,28 +43,28 @@ export class TasksRepo {
   }
 
   /**
- * Delete task by id
- * @param id - task id string
- */
+   * Delete task by id
+   * @param id - task id string
+   */
   async deleteOne(id: string) {
     tasks = tasks.filter((elem) => elem.id !== id);
   }
 
   /**
- * Delete all filtered tasks
- * @param filter - object with key to search and value to compare
- */
+   * Delete all filtered tasks
+   * @param filter - object with key to search and value to compare
+   */
   async deleteMany({ key, value }: { key: keyof TaskClass, value: string }) {
     const filteredCollection = tasks.filter((elem) => elem[key] !== value);
     tasks = filteredCollection;
   }
 
   /**
- * Update task by id
- * @param id - task id string
- * @param itemData - data to update a particular task
- * @returns \{Promise\} Promise object represents updated task
- */
+   * Update task by id
+   * @param id - task id string
+   * @param itemData - data to update a particular task
+   * @returns \{Promise\} Promise object represents updated task
+   */
   async updateOne(id: string, itemData: TaskClass) {
     const itemIndex = tasks.findIndex((elem) => elem.id === id);
 
@@ -74,10 +74,10 @@ export class TasksRepo {
   }
 
   /**
- * Update many tasks found using filter
- * @param filter - object with key to search and value to compare
- * @param updates - data to update each task
- */
+   * Update many tasks found using filter
+   * @param filter - object with key to search and value to compare
+   * @param updates - data to update each task
+   */
   async updateMany(filter: { key: keyof TaskClass, value: string }, updates: Partial<TaskClass>) {
     const filteredItems = tasks.filter((elem: TaskClass) => elem[filter.key] === filter.value);
     const filteredItemsIndexes = filteredItems.map((elem) => tasks.findIndex((collectionItem) => collectionItem.id === elem.id));
